+++ app/actions/viewerActions.js
@@ -7,6 +7,7 @@ export const POP_EDITSTACK = 'POP_EDITSTACK';
 export const RESET_EDITSTACK = 'RESET_EDITSTACK';
 export const SAVE_RSML = 'SAVE_RSML';
 export const UPDATE_CHECKED = 'CHECK_FOLDER';
+export const UPDATE_VIEWER_FILTER = 'UPDATE_VIEWER_FILTER';
 
 export const addViewer      = viewerID => ({ type: ADD_VIEWER, viewerID}); //ID obtained by process.pid
 export const removeViewer   = viewerID => ({ type: REMOVE_VIEWER, viewerID });
@@ -16,4 +17,5 @@ export const pushEditStack  = (viewerID, lines) => ({ type: PUSH_EDITSTACK, view
 export const popEditStack = viewerID => ({ type: POP_EDITSTACK, viewerID });
 export const resetEditStack = viewerID => ({ type: RESET_EDITSTACK, viewerID });
 export const saveRSML = (viewerID, newRSML) => ({ type: SAVE_RSML, viewerID, newRSML });
-export const updateChecked = (viewerID, checked) => ({ type: UPDATE_CHECKED, viewerID, checked })
+export const updateChecked = (viewerID, checked) => ({ type: UPDATE_CHECKED, viewerID, checked });
+export const updateViewerFilter = (viewerID, text) => ({ type: UPDATE_VIEWER_FILTER, viewerID, text });