+++ app/components/gallery/FilterBar.js
@@ -25,7 +25,7 @@ export default class FilterBar extends Component{
     {
         if (this.typingTimeout) clearTimeout(this.typingTimeout);
         this.text = e.target.value,
-        this.typingTimeout = setTimeout(() => this.props.updateFilterText(this.text), 200);
+        this.typingTimeout = setTimeout(() => this.props.updateFilterText(this.text.toLowerCase()), 200);
     }
 
     updateFilterAnalysed = e =>
@@ -44,21 +44,21 @@ export default class FilterBar extends Component{
         return (
             <StyledFilterBarSpan>
                 <InputGroup>
-                <InputGroup.Prepend>
-                    <InputGroup.Text><StyledIcon className={"fas fa-search fa-lg"}/></InputGroup.Text> 
-                </InputGroup.Prepend>
-                <input key={0} type="text" className="form-control" placeholder="Filter images..." onChange={this.updateFilterText} ref={this.textref}/>
-                <button key={1} className="btn bg-transparent" style={{'marginLeft': '-40px', 'zIndex': '100'}} onClick={this.clear}>
-                    <i className="fa fa-times"></i>
-                </button>
-                <InputGroup.Append>
-                    <InputGroup.Text>
-                    <div className="custom-control custom-checkbox">
-                        <input type="checkbox" className="custom-control-input" id="analysed" onClick={this.updateFilterAnalysed} ref={this.checkboxref}/>
-                        <label className="custom-control-label" htmlFor="analysed">Analysed</label>
-                    </div>
-                    </InputGroup.Text>
-                </InputGroup.Append>
+                    <InputGroup.Prepend>
+                        <InputGroup.Text><StyledIcon className={"fas fa-search fa-lg"}/></InputGroup.Text> 
+                    </InputGroup.Prepend>
+                    <input key={0} type="text" className="form-control" placeholder="Filter images..." onChange={this.updateFilterText} ref={this.textref}/>
+                    <button key={1} className="btn bg-transparent" style={{'marginLeft': '-40px', 'zIndex': '100'}} onClick={this.clear}>
+                        <i className="fa fa-times"></i>
+                    </button>
+                    <InputGroup.Append>
+                        <InputGroup.Text>
+                        <div className="custom-control custom-checkbox">
+                            <input type="checkbox" className="custom-control-input" id="analysed" onClick={this.updateFilterAnalysed} ref={this.checkboxref}/>
+                            <label className="custom-control-label" htmlFor="analysed">Analysed</label>
+                        </div>
+                        </InputGroup.Text>
+                    </InputGroup.Append>
                 </InputGroup>
             </StyledFilterBarSpan>
         );